requirements.txt
#Langchain Core
langchain 
langchain-core

#OpenAI iNTEGRATION
langchain-openai
openai
langchain-community

#Anthropic Integration
langchain-anthropic

#Google Gemini Integration
langchain-google-genai
google-generativeai

#Hugging Face Integration
langchain-huggingface
transformers
huggingface-hub

#Environment Variables
python-dotenv

#Machine Learning
scikit-learn
numpy


# Additional Libraries
langchain
langchain-huggingface
sentence-transformers
